+++ src/db/connect.ts
@@ -1,16 +1,16 @@
-import logger from "../logger";
-import mongoose from "mongoose";
+import logger from '../logger';
+import mongoose from 'mongoose';
 
 const connect = async () => {
   const url: string = process.env.DB_URI as string;
   mongoose.connect(url);
   mongoose.pluralize(null);
-  mongoose.connection.once("open", () => {
-    logger.info("Connected to database");
+  mongoose.connection.once('open', () => {
+    logger.info('Connected to MongoDB');
   });
 
-  mongoose.connection.on("error", (err) => {
-    logger.error("Error connecting to database...", err);
+  mongoose.connection.on('error', err => {
+    logger.error('Error connecting to MongoDB...', err);
     process.exit();
   });
 };
@@ -20,8 +20,8 @@ const disconnect = () => {
     return;
   }
   mongoose.disconnect();
-  mongoose.connection.once("close", async () => {
-    console.log("Diconnected  to database");
+  mongoose.connection.once('close', async () => {
+    console.log('Diconnected to database');
   });
 };