+++ src/db/redis.ts
@@ -1,26 +1,26 @@
 import logger from '../logger';
-
 import { createClient } from 'redis';
 import { scrapeMangaList } from '../services/manga.services';
 
 export const connectToRedis = async () => {
-  try {
-    const redisUri = process.env.REDIS_URI as string;
-    const redisClient = createClient({ url: redisUri });
+    const redisHost = process.env.REDIS_HOST as string;
+    const redisPort = process.env.REDIS_PORT as string;
+    const redisPassword = process.env.REDIS_KEY as string;
+    const redisClient = createClient({ url: `redis://default:${redisPassword}@${redisHost}:${redisPort}` });
     await redisClient.connect();
     redisClient.on('error', err => logger.error('Redis Client Error', err));
+    console.log('Resting Manga List in Redis DB')
     const list = await scrapeMangaList();
     redisClient.SET('mangaList', JSON.stringify(list));
     logger.info('mangaList updated in Redis!');
-  } catch (error) {
-    logger.error('Error connecting to database...', error);
-  }
 };
 
 export const getMangaList = async () => {
   try {
-    const redisUri = process.env.REDIS_URI as string;
-    const redisClient = createClient({ url: redisUri });
+    const redisHost = process.env.REDIS_HOST as string;
+    const redisPort = process.env.REDIS_PORT as string;
+    const redisPassword = process.env.REDIS_KEY as string;
+    const redisClient = createClient({ url: `redis://default:${redisPassword}@${redisHost}:${redisPort}` });
     await redisClient.connect();
     redisClient.on('error', err => logger.error('Redis Client Error', err));
     const mangaList = (await redisClient.GET('mangaList')) as string;